pre weekend

+++ src/minecrunch.models/Chunks/Chunk.cs
@@ -13,7 +13,7 @@ namespace minecrunch.models.Chunks
         public int y;
         public Biome biome;
         public ChunkSection[] sections = new ChunkSection[16];
-        public int[,] SurfaceMap = new int[16, 16];
+        public double[,] SurfaceMap = new double[16, 16];
         public long processTimeMs = 0;
 
         public Block GetBlockByChunkCoord(int x, int y, int z)
